requirements.txt
SQLAlchemy
pymysql
flask
PyJWT
pycryptodome
python-dotenv